more control over std.ignore in dl_read_gcp
the changes here do not impact how dl_read_gcp works, but do allow
windows users to change the ignore.std inputs for the cmd function. I do
now know why it happens, but cmd() can throw an error in local R
sessions for windows users when both ignore.stderr and ignore.stdout are
TRUE. I added documentation explaining the new inputs and a note in the
details about the issue and how to fix it.